Used in my genomics course
Here is the progression of this code we used to build a fly mapping simulator. The idea here is that the code will give data for certain conditions of the fly's genotype. The purpose of this is to use as a learning software. The professor is able to generate many data sets and will know the answer key and the student must work backwards to understand the map units and allele orders.